// src/types/salesSummary.ts
export default interface SalesSummary {
    "Vbeln": string;
    "Audat": string;
    "Kunnr": string;
    "Vkorg": string;
    "Vtweg": string;
    "Spart": string;
    "Netwr": string;
    "Waerk": string;
    "Vkbur": string;
    "Vkgrp": string;
    "Bstnk": string;
    "Matnr": string;
    "Arktx": string;
    "Kwmeng": string;
    "Vrkme": string;
    "Pstyv": string;
    "Posex": string;
}